teleport/requirements.txt
amqp==2.5.2
asgiref==3.2.7
beautifulsoup4==4.11.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.2
celery-stubs==0.1.1
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
distro==1.5.0
dj-database-url==0.5.0
Django==3.0.5
django-appconf==1.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
instagram-private-api==1.6.0.0
kombu==4.6.8
lxml==4.5.0
mypy==0.942
mypy-extensions==0.4.3
numpy==1.21.5
openpyxl==3.0.10
pandas==1.3.5
Pillow==7.1.2
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.3
pytils==0.3
pytz==2019.3
redis==3.4.1
requests==2.23.0
selection==0.0.14
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tomli==2.0.1
typed-ast==1.5.2
typing-extensions==4.1.1
urllib3==1.25.9
user-agent==0.1.9
uuid==1.30
vine==1.3.0
weblib==0.1.30
whitenoise==5.0.1
zipp==3.1.0
